Allowing duplicate definitions when parsing dsdl
duplicate == the same thing more than once. This is okay.
redefined == the same identifer used by incompatible definitions.
this is not okay.

In complex build systems duplicate dsdl files can get pulled in from
multiple dependencies and end up in long lists of paths that are
passed into the pyuavcan dsdl compiler. This change to the dsdl
parser allows this without throwing errors but enforces that the
duplicated definitions are compatible. This elminiates the need
for special case handling in build systems to try to find the duplicates
before invoking pyuavcan and adds a new check where such complex
systems will be notified when duplicated definitions get out of
sync.