Updated SocketcandMedia class with unit test
Added unit test and changed spec string to be parsed in _transport_factory

fixed formatting issues that did not pass Black guidelines

Removed unecessary params S

Updated version of the socketcandMedia class. This fixes the unnecessary pieces left from the pythonCANMedia class, such as the FD vs Classic dataclasses, Can Bus Options, and the constructor class. It also updates the way parameters are inputed to be all in the iface_name param like 'socketcand:can:host:port'

SocketcandMedia Media class added for socketcand library functionality. It runs on Python-CAN like the interfaces in the PythonCANMedia class, but because it has extra variables (host and port), it is not very clean to just add socketcand to PythonCANMedia.