Updated interface assert and fixed infinite test failure
In the previous commit it was mostly doc updates, but one changed the
socketcandmedia interface_name return. I forgot to update the unit test
assertions to match this, so in the CI testing it resulted in a
never ending loop. The issue stemmed from the way Popen gets stderr
and stdout with the communicate method, but only after the yield
statement. It does not respond to the kill() and instead runs forever,
so I replaced it with just a kill. Although you lose a little bit of
info, the socketcand instance gets checked before the unit test
is done in the fixture, and from there the daemon is very robust and
will not stop unless you force kill it or terminate it. So I think this
is the best solution, you still get output from errors within the unit
test as well.